// rs-clone/client/src/enums/app-tag.ts
const enum AppTag {
  HEADER = 'header',
  DIV = 'div',
  SPAN = 'span',
  BUTTON = 'button',
  LINK = 'a',
  H1 = 'h1',
  H2 = "h2",
  H3 = "h3",
  NAV = 'nav',
  P = 'p',
  SECTION = 'section',
  FORM = 'form',
  UL = "ul",
  LI = "li",
  A = "a",
  INPUT = 'input',
  LABEL = 'label',
  MAIN = 'main',
}

export default AppTag;